feat: update timecopilot results based on dashboard enhacements (#45)
first of all, thank you for the recent updates to the GIFT‑Eval
dashboard! :)

the new data‑leakage flag is a huge help for both practitioners and
researchers, it makes it much easier to identify models that may have
seen parts of the test set during training. likewise, thank you for
adding the agentic flag, which is another excellent addition!

to keep TimeCopilot's results focused on truly unseen data, we’ve
excluded any models flagged for potential data leakage from the original
experiment and added the Sundial model (a recent addition to the
library) to the ensemble. the config file has also been updated to
reflect that we are now only using models not flagged for data leakage.
the notebook has been updated accordingly, with a note at the end
explaining these changes in detail and referencing the original
experiment for comparison.

the repo was also updated to include this changelog, you can find it
[here](https://github.com/AzulGarza/timecopilot/tree/main/experiments/gift-eval#changelog).

thank you for maintaining this repo and continuing to improve such a
valuable resource!