Add evaluation results on TSOrchestra-test (#51)
Hi,

We would like to submit our new ensemble results, named
'TSOrchestra-test', to the GiftEval leaderboard. As discussed in PR
https://github.com/SalesforceAIResearch/gift-eval/pull/27 and
https://github.com/SalesforceAIResearch/gift-eval/pull/44. This is the
ensemble architecture with _**'tabpfn_ts', 'yinglong_300m', 'Tempo',
'toto', 'tirex'**_, we are continuing to work on improving the results,
so we will upload the evaluation pipeline and our new ensemble-based FM
works together. The target is to submit them in mid-to-late September.
Thank you for your consideration!

---------

Co-authored-by: Taha Aksu <tahaaksu01@gmail.com>