Add Nexus-1.0 results to the leaderboard (#58)
Hi,
Thank you so much for your wonderful work on releasing GIFT-Eval and for
maintaining it. We would like to submit evaluation results with our new
method "Nexus-1.0". Nexus is an agentic work in progress. In this
submitted version of results, Nexus is a sophisticated ensemble
approach, on which we will be publishing the detailed paper very soon.
We will also be working on improving the results.

Thank you so much once again!