Include Kairos-1.0 results in the leaderboard. (#60)
Hello,
Thank you for the outstanding work in building and maintaining
GIFT-Eval. We would like to contribute results from our model Kairos.
Kairos leverages Toto, Flowstate, and TimesFM2.5 as its backbone within
an ensemble framework, carefully evaluated with no leakage.
 
Within the scope of our technology, Kairos is enriched with agentic
behavior: it can provide natural language explanations automatically and
in real time whenever interesting areas in the stream are detected.
Further details will be presented in the forthcoming white paper on [our
technology](https://continualist.ai/).
 
We are grateful for the opportunity to contribute to this benchmark and
look forward to sharing more soon.
Thanks.

Gere