add the Xihe-max results (#73)
Hello,
We would like to submit results from our model, Xihe-max, to the
leaderboard. Additional details will be provided in our paper, which is
currently under review. We appreciate the opportunity to contribute to
this benchmark.
We have updated the config file following a recent new column addition
to the leaderboard regarding replication code.
Thank you!